Move PublicRegisterView from danceschool.register to danceschool.core (Step 2)
Uses git mv on register/views.py → core/registration.py to preserve history,
then restores register/views.py containing only PointOfSaleRegisterView.

Changes:
- danceschool/core/registration.py: new home for PublicRegisterView; all
  imports are now relative to core (no register-app dependencies)
- danceschool/register/views.py: recreated with only PointOfSaleRegisterView
- danceschool/core/urls_registration.py: adds publicRegistration,
  publicRegistrationWithVoucher, publicRegistrationWithMarketingId URL patterns
- danceschool/register/urls.py: public-register URLs removed (now in core);
  retains only registerView and registerAutocomplete patterns
- core/templates/core/public_register.html: moved from
  register/templates/register/public_register.html via git mv
- PublicRegisterView.template_name updated to 'core/public_register.html'
