+++ danceschool/core/templates/core/public_register.html
